reformat

+++ website/docs/guides/fixtures/books.js
@@ -26,12 +26,12 @@ db.books.insertMany([
       { country: 'India', format: ['E-book'], released: true, release_date: '2010-01-01T00:00:00Z' }
     ],
     isbn: '978-0141439518',
-    pages: 448,
+    pages: Int32(448),
     price: { value: 19.99, currency: 'USD' },
     keywords: ['Elizabeth Bennet', 'Mr. Darcy', '19th Century'],
     reviews: [
-      { user: 'johnny_moe', rating: 5, comment: 'An all-time favorite!' },
-      { user: 'jane_smith', rating: 4, comment: 'A classic tale with timeless lessons.' }
+      { user: 'johnny_moe', rating: Int32(5), comment: 'An all-time favorite!' },
+      { user: 'jane_smith', rating: Int32(4), comment: 'A classic tale with timeless lessons.' }
     ],
     textual_analysis: {
       description: 'A story of love and societal expectations in 19th-century England.',
@@ -39,8 +39,8 @@ db.books.insertMany([
     },
     analytics: {
       average_rating: 4.5,
-      ratings_count: 1500,
-      sales: { units_sold: 1000000, countries: ['United States', 'United Kingdom', 'India'] }
+      ratings_count: Int32(1500),
+      sales: { units_sold: Int32(1000000), countries: ['United States', 'United Kingdom', 'India'] }
     },
     digital_metadata: { download_link: 'https://www.gutenberg.org/ebooks/1342', formats: ['PDF', 'EPUB', 'MOBI'] },
     summary:
@@ -72,18 +72,18 @@ db.books.insertMany([
       }
     ],
     isbn: '978-1503280786',
-    pages: 378,
+    pages: Int32(378),
     price: { value: 15.99, currency: 'USD' },
     keywords: ['Captain Ahab', 'Whaling', 'Revenge'],
-    reviews: [{ user: 'sailor_sam', rating: 5, comment: 'A thrilling adventure on the high seas.' }],
+    reviews: [{ user: 'sailor_sam', rating: Int32(5), comment: 'A thrilling adventure on the high seas.' }],
     textual_analysis: {
       description: 'A thrilling tale of obsession and revenge at sea.',
       themes: ['Obsession', 'Revenge', 'Fate']
     },
     analytics: {
       average_rating: 4.3,
-      ratings_count: 800,
-      sales: { units_sold: 500000, countries: ['United States', 'United Kingdom'] }
+      ratings_count: Int32(800),
+      sales: { units_sold: Int32(500000), countries: ['United States', 'United Kingdom'] }
     },
     digital_metadata: { download_link: 'https://www.gutenberg.org/ebooks/2701', formats: ['PDF', 'EPUB', 'MOBI'] },
     summary:
@@ -115,18 +115,20 @@ db.books.insertMany([
       }
     ],
     isbn: '978-0199537150',
-    pages: 336,
+    pages: Int32(336),
     price: { value: 10.99, currency: 'USD' },
     keywords: ['Victor Frankenstein', 'Monster', 'Creation'],
-    reviews: [{ user: 'science_geek', rating: 5, comment: 'A chilling exploration of ambition and consequence.' }],
+    reviews: [
+      { user: 'science_geek', rating: Int32(5), comment: 'A chilling exploration of ambition and consequence.' }
+    ],
     textual_analysis: {
       description: 'A dark tale of scientific ambition and its unintended consequences.',
       themes: ['Ambition', 'Morality', 'Isolation']
     },
     analytics: {
       average_rating: 4.7,
-      ratings_count: 1200,
-      sales: { units_sold: 600000, countries: ['United Kingdom', 'United States'] }
+      ratings_count: Int32(1200),
+      sales: { units_sold: Int32(600000), countries: ['United Kingdom', 'United States'] }
     },
     digital_metadata: { download_link: 'https://www.gutenberg.org/ebooks/84', formats: ['PDF', 'EPUB', 'MOBI'] },
     summary:
@@ -162,12 +164,12 @@ db.books.insertMany([
       }
     ],
     isbn: '978-0131873254',
-    pages: 1119,
+    pages: Int32(1119),
     price: { value: 175.99, currency: 'USD' },
     keywords: ['Database Design', 'SQL', 'Data Storage', 'Query Processing'],
     reviews: [
-      { user: 'tech_reader', rating: 5, comment: 'Comprehensive resource for database systems.' },
-      { user: 'db_enthusiast', rating: 4, comment: 'In-depth coverage but quite dense.' }
+      { user: 'tech_reader', rating: Int32(5), comment: 'Comprehensive resource for database systems.' },
+      { user: 'db_enthusiast', rating: Int32(4), comment: 'In-depth coverage but quite dense.' }
     ],
     textual_analysis: {
       description:
@@ -176,8 +178,8 @@ db.books.insertMany([
     },
     analytics: {
       average_rating: 4.5,
-      ratings_count: 100,
-      sales: { units_sold: 50000, countries: ['United States', 'Canada', 'United Kingdom'] }
+      ratings_count: Int32(100),
+      sales: { units_sold: Int32(50000), countries: ['United States', 'Canada', 'United Kingdom'] }
     },
     digital_metadata: { download_link: null, formats: ['PDF', 'EPUB'] },
     summary:
@@ -209,14 +211,14 @@ db.books.insertMany([
       }
     ],
     isbn: '978-0132350884',
-    pages: 464,
+    pages: Int32(464),
     price: { value: 49.99, currency: 'USD' },
     keywords: ['Software Craftsmanship', 'Code Quality', 'Agile Development', 'Programming Best Practices'],
     reviews: [
-      { user: 'dev_guru', rating: 5, comment: 'An essential guide for writing clean, maintainable code.' },
+      { user: 'dev_guru', rating: Int32(5), comment: 'An essential guide for writing clean, maintainable code.' },
       {
         user: 'code_master',
-        rating: 4,
+        rating: Int32(4),
         comment: 'Provides practical advice, though some examples are language-specific.'
       }
     ],
@@ -227,8 +229,8 @@ db.books.insertMany([
     },
     analytics: {
       average_rating: 4.7,
-      ratings_count: 2500,
-      sales: { units_sold: 150000, countries: ['United States', 'Canada', 'United Kingdom'] }
+      ratings_count: Int32(2500),
+      sales: { units_sold: Int32(150000), countries: ['United States', 'Canada', 'United Kingdom'] }
     },
     digital_metadata: { download_link: null, formats: ['PDF', 'EPUB'] },
     summary:
