+++ website/docs/guides/requests/ttl-indexes/1-create-ttl-index.js
@@ -1,4 +1,4 @@
 db.runCommand({
   createIndexes: 'books',
-  indexes: [{ key: { 'reservation.date': 1 }, name: 'reservation_ttl', expireAfterSeconds: 60 }]
+  indexes: [{ key: { 'reservation.date': Int32(1) }, name: 'reservation_ttl', expireAfterSeconds: Int32(60) }]
 })
